% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_cla_nb.R
\name{cla_nb}
\alias{cla_nb}
\title{Gaussian Naive Bayes Classifier}
\usage{
cla_nb(attribute, slevels, var_smoothing = 1e-09, priors = NULL)
}
\arguments{
\item{attribute}{Target attribute name for model building}

\item{slevels}{List of possible values for classification target}

\item{var_smoothing}{Portion of the largest variance of all features that is added to variances}

\item{priors}{Prior probabilities of the classes. If specified must be a list of length n_classes}
}
\value{
A Naive Bayes classifier object

\code{cla_nb} object
}
\description{
Implements classification using the Gaussian Naive Bayes algorithm.
This function wraps the GaussianNB from Python's scikit-learn library.
}
\details{
Naive Bayes Classifier
}
\examples{
#See an example of using `cla_nb` at this
#https://github.com/cefet-rj-dal/daltoolboxdp/blob/main/examples/cla_nb.md
}
