% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_selectf_model_lg.R
\name{create_fe_lg_model}
\alias{create_fe_lg_model}
\alias{create_fe_selectfrommodel_lg}
\alias{fit_transform_fe_selectfrommodel_lg}
\title{SelectFromModel + Logistic Regression}
\usage{
create_fe_lg_model(
  df_train,
  target_column,
  C = 0.1,
  penalty = "l1",
  solver = "liblinear"
)

create_fe_selectfrommodel_lg(model, threshold = "mean", prefit = TRUE)

fit_transform_fe_selectfrommodel_lg(select_method, df_train, target_column)
}
\arguments{
\item{df_train}{Data frame to transform}

\item{target_column}{The target column name as string}

\item{C}{Inverse of regularization strength}

\item{penalty}{Norm used in the penalization ("l1", "l2", etc.)}

\item{solver}{Algorithm to use in optimization ("liblinear", …)}

\item{model}{A prefit Python model}

\item{threshold}{Threshold for feature selection (e.g. "mean")}

\item{prefit}{Logical; whether the model is already fitted}

\item{select_method}{The feature selection model (Python object)}
}
\value{
A fitted Python Logistic Regression model

A Python SelectFromModel object

Transformed X_train with selected features
}
\description{
Apply SelectFromModel using a LogisticRegression estimator.
}
\section{Functions}{
\itemize{
\item \code{create_fe_selectfrommodel_lg()}: Create a feature selector using SelectFromModel

\item \code{fit_transform_fe_selectfrommodel_lg()}: Fit and transform using SelectFromModel LG

}}
