% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_selectf_model_rf.R
\name{create_fe_rf_model}
\alias{create_fe_rf_model}
\alias{create_fe_selectfrommodel_rf}
\alias{fit_transform_fe_selectfrommodel_rf}
\title{Feature Selection Using SelectFromModel and RandomForest}
\usage{
create_fe_rf_model(
  df_train,
  target_column,
  n_estimators = 100,
  random_state = 0
)

create_fe_selectfrommodel_rf(model, threshold = "mean", prefit = TRUE)

fit_transform_fe_selectfrommodel_rf(select_method, df_train, target_column)
}
\arguments{
\item{df_train}{Data frame to transform}

\item{target_column}{The target column name as string}

\item{n_estimators}{number estimators}

\item{random_state}{random_state sets the seed for the random number generator.}

\item{model}{A Python Logistic Regression model}

\item{threshold}{The threshold for feature selection}

\item{prefit}{Boolean indicating if the model should be considered prefit}

\item{select_method}{The feature selection model (Python object)}
}
\value{
A fitted Python RandomForest model

A Python SelectFromModel object

Transformed X_train with selected features
}
\description{
This module applies feature selection using SelectFromModel with a RandomForestClassifier estimator.
}
\section{Functions}{
\itemize{
\item \code{create_fe_selectfrommodel_rf()}: Create a feature selector using SelectFromModel

\item \code{fit_transform_fe_selectfrommodel_rf()}: Fit and transform using SelectFromModel RF

}}
