% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_select_rfe.R
\name{create_fe_rfe_model}
\alias{create_fe_rfe_model}
\alias{fit_transform_fe_rfe}
\title{Feature Selection Using RFE and Logistic Regression}
\usage{
create_fe_rfe_model(n_features_to_select = 0.5, lg_max_iter = 1000)

fit_transform_fe_rfe(select_method, df_train, target_column)
}
\arguments{
\item{n_features_to_select}{Number or fraction of features to select}

\item{lg_max_iter}{Maximum number of iterations for the LogisticRegression estimator}

\item{select_method}{The RFE model (Python object)}

\item{df_train}{Data frame to transform}

\item{target_column}{The target column name as string}
}
\value{
A Python RFE object

Transformed X_train with selected features
}
\description{
This module applies Recursive Feature Elimination (RFE) with a LogisticRegression estimator.
}
\section{Functions}{
\itemize{
\item \code{fit_transform_fe_rfe()}: Fit and transform using RFE

}}
