% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_selectk_beast.R
\name{create_fe_selectkbest_model}
\alias{create_fe_selectkbest_model}
\alias{fit_transform_fe_selectkbest}
\title{Feature Selection Using SelectKBest}
\usage{
create_fe_selectkbest_model(k = 10)

fit_transform_fe_selectkbest(select_method, df_train, target_column)
}
\arguments{
\item{k}{Number of top features to select}

\item{select_method}{The SelectKBest model (Python object)}

\item{df_train}{Data frame to transform}

\item{target_column}{The target column name as string}
}
\value{
A Python SelectKBest object

Transformed X_train with selected features
}
\description{
Applies feature selection using SelectKBest from scikit-learn.
}
