% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_select_variance_threshold.R
\name{create_fe_variance_threshold_model}
\alias{create_fe_variance_threshold_model}
\title{Feature Selection Using VarianceThreshold}
\usage{
create_fe_variance_threshold_model(threshold = 0.2)
}
\arguments{
\item{threshold}{The variance threshold below which features will be removed}
}
\value{
A Python VarianceThreshold object
}
\description{
Applies feature selection based on variance using scikit-learn's VarianceThreshold.
}
