% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_select_sequential.R
\name{fit_transform_fe_sequential_fs}
\alias{fit_transform_fe_sequential_fs}
\title{Fit and Transform with Sequential Feature Selector}
\usage{
fit_transform_fe_sequential_fs(select_method, df_train, target_column)
}
\arguments{
\item{select_method}{A SequentialFeatureSelector Python object created by \code{create_fe_sequential_fs_model()}.}

\item{df_train}{An R data frame of training data.}

\item{target_column}{Character name of the target column in \code{df_train}.}
}
\value{
A data frame containing only the features selected by the SequentialFeatureSelector.
}
\description{
Apply a pre‐configured SequentialFeatureSelector (Python) on a data frame and return only the selected features.
}
