% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_ig.R
\name{fs_ig}
\alias{fs_ig}
\title{Information Gain}
\usage{
fs_ig(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
A \code{fs_ig} object.
}
\description{
Information Gain is a feature selection technique based on information theory. It measures the information obtained for the target variable by knowing the presence or absence of a feature.
It wraps the FSelector library.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(fs_ig("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
