% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skcla_gb.R
\name{skcla_gb}
\alias{skcla_gb}
\title{Gradient Boosting Classifier}
\usage{
skcla_gb(
  attribute,
  slevels,
  n_estimators = 100,
  learning_rate = 0.1,
  max_depth = 3,
  subsample = 1,
  min_samples_split = 2,
  min_samples_leaf = 1,
  loss = c("log_loss", "exponential")
)
}
\arguments{
\item{attribute}{Target attribute name for model building.}

\item{slevels}{Possible values for the target classification.}

\item{n_estimators}{Number of boosting stages to perform.}

\item{learning_rate}{Learning rate that shrinks the contribution of each tree.}

\item{max_depth}{Maximum depth of the individual regression estimators.}

\item{subsample}{Fraction of samples used to fit each stage.}

\item{min_samples_split}{Minimum number of samples required to split an internal node.}

\item{min_samples_leaf}{Minimum number of samples required to be at a leaf node.}

\item{loss}{Loss function to be optimized. One of \code{"log_loss"} or \code{"exponential"}.}
}
\value{
A \code{skcla_gb} classifier object.
}
\description{
Implements a classifier using the Gradient Boosting algorithm.
Wraps scikit-learn's \code{GradientBoostingClassifier} through \code{reticulate}.
}
\details{
Tree Boosting
}
\examples{
\dontrun{
data(iris)
clf <- skcla_gb(
  attribute = "Species",
  slevels = levels(iris$Species),
  n_estimators = 150,
  learning_rate = 0.05
)
clf <- daltoolbox::fit(clf, iris)
pred <- predict(clf, iris)
table(pred, iris$Species)
}
}
\references{
Friedman, J. H. (2001). Greedy Function Approximation: A Gradient Boosting Machine.
}
