% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skcla_svc.R
\name{skcla_svc}
\alias{skcla_svc}
\title{Support Vector Machine Classification}
\usage{
skcla_svc(
  attribute,
  slevels,
  C = 1,
  kernel = c("rbf", "linear", "poly", "sigmoid"),
  gamma = "scale",
  degree = 3,
  coef0 = 0,
  probability = FALSE,
  class_weight = NULL
)
}
\arguments{
\item{attribute}{Target attribute name for model building.}

\item{slevels}{List of possible values for classification target.}

\item{C}{Regularization strength parameter.}

\item{kernel}{Kernel function type. One of \code{"rbf"}, \code{"linear"}, \code{"poly"}, or \code{"sigmoid"}.}

\item{gamma}{Kernel coefficient value. Use \code{"scale"}, \code{"auto"}, or a numeric value.}

\item{degree}{Polynomial degree when using \code{kernel = "poly"}.}

\item{coef0}{Independent term value in polynomial and sigmoid kernels.}

\item{probability}{Whether to enable probability estimates.}

\item{class_weight}{Optional weights associated with classes.}
}
\value{
A \code{skcla_svc} classifier object.
}
\description{
Implements classification using support vector machines.
Wraps scikit-learn's \code{SVC} through \code{reticulate}.
}
\details{
SVM Classifier
}
\examples{
\dontrun{
data(iris)
clf <- skcla_svc(
  attribute = "Species",
  slevels = levels(iris$Species),
  kernel = "rbf",
  C = 1
)
clf <- daltoolbox::fit(clf, iris)
pred <- predict(clf, iris)
table(pred, iris$Species)
}
}
\references{
Cortes, C., & Vapnik, V. (1995). Support-Vector Networks.
}
