% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_select_sequential.R
\name{create_fe_sequential_fs_model}
\alias{create_fe_sequential_fs_model}
\title{Feature Selection Using SequentialFeatureSelector}
\usage{
create_fe_sequential_fs_model(
  n_neighbors = 3,
  direction = "forward",
  n_features_to_select = 2
)
}
\arguments{
\item{n_neighbors}{Number of neighbors for KNN}

\item{direction}{Direction of selection: "forward" or "backward"}

\item{n_features_to_select}{Number of features to select}
}
\value{
A Python SequentialFeatureSelector object
}
\description{
This module applies sequential feature selection using KNeighborsClassifier.
}
