% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fss.R
\name{fs_fss}
\alias{fs_fss}
\title{Forward Stepwise Selection}
\usage{
fs_fss(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
A \code{fs_fss} object.
}
\description{
Forward stepwise selection is a technique for feature selection in which attributes are added to a model one at a time based on their ability to improve the model's performance. It stops adding once the candidate addition does not significantly improve model adjustment.
It wraps the leaps library.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(fs_fss("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
