% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_relief.R
\name{fs_relief}
\alias{fs_relief}
\title{Relief}
\usage{
fs_relief(attribute)
}
\arguments{
\item{attribute}{The target variable.}
}
\value{
A \code{fs_relief} object.
}
\description{
Feature selection using Relief is a technique for selecting a subset of relevant features. It calculates the relevance of a feature by considering the difference in feature values between nearest neighbors of the same and different classes.
It wraps the FSelector library.
}
\examples{
data(iris)
myfeature <- daltoolbox::fit(fs_relief("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
}
